examples/agno_ex/content_agent_noframework/requirements.txt
gradio>=4.0.0
jinja2>=3.0.0
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.28.0
webbrowser
pathlib
agno>=0.1.0
firecrawl-py>=0.1.0 